requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sentence-transformers==2.2.2
torch>=2.0.0
